reorganize the test cases
