+++ LambdaAndJdkCollections/src/test/java/org/sample/lambda/AllTests.java
@@ -1,0 +1,11 @@
+package org.sample.lambda;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+@RunWith(Suite.class)
+@SuiteClasses({ LoopElementTest.class, RemoveElementTest.class })
+public class AllTests {
+
+}
